requirements.txt
asgiref==3.8.1
cffi==1.17.1
dj-database-url==2.3.0
Django==5.1.7
djangorestframework==3.15.2
geographiclib==2.0
geopy==2.4.1
h3==4.2.2
numpy==2.2.4
pillow==11.1.0
psycopg2==2.9.10
pycparser==2.22
python-dotenv==1.0.1
pytz==2025.2
sqlparse==0.5.3
timezonefinder==6.5.9
typing_extensions==4.12.2
tzdata==2025.1
